+++ lib/stds/stds.c
@@ -46,13 +46,13 @@ void calcSlope(int x1, int y1, int x2, int y2, float* dx, float* dy) {
 }
 
 float getAngle(int32_t x1, int32_t y1, int32_t x2, int32_t y2) {
-  float angle = -90 + atan2(y1 - y2, x1 - x2) * (180 / PI);
+  float angle = (float) (-90 + atan2(y1 - y2, x1 - x2) * (180 / PI));
   return angle >= 0 ? angle : 360 + angle;
 }
 
-int getDistance(int32_t x1, int32_t y1, int32_t x2, int32_t y2) {
+int32_t getDistance(int32_t x1, int32_t y1, int32_t x2, int32_t y2) {
   int x = x2 - x1;
   int y = y2 - y1;
 
-  return sqrt(x * x + y * y);
+  return (int32_t) sqrt(x * x + y * y);
 }
